backend/requirements.txt
uvicorn
fastapi
python-dotenv
requests
pydantic